Extract run_maker startup into startup::run_startup (#255)
The first ~410 lines of run_maker (arg validation, symbol/config
resolution, and the live clean-start handshake) were pure setup that
either fails fast or produces a fixed set of values the quoting loop
consumes read-only. Move them into a new startup.rs behind
`run_startup(symbol, &args, output_format) -> Result<MakerStartup>`.

- `MakerStartup` bundles the 10 outputs the loop needs (client, cfg,
  canonical symbol, notifier, qty_tolerance, run_order_prefix, the
  adopted ledger baseline, and the Option<LiveSession>). run_maker now
  destructures it in one statement.
- `LiveSession` moves to startup.rs as the "session structure" it
  produces; fields become pub(super) so the loop keeps its plain-field
  access. `new_maker_rest_client` and `validate_alert_thresholds` move
  with it (both are startup concerns); their tests follow via re-export.
- run_maker drops from ~2263 to ~1867 lines. Pure mechanical move: no
  control-flow or ordering change, covered by the existing behavior
  tests plus a paper smoke run (quote cycles + clean Ctrl+C) and a
  fail-fast arg-validation check.

Part of #255 (the freeze->cleanup->recover skeleton merge is a separate
follow-up).
