+++ crates/standx-cli/src/commands/maker/runtime.rs
@@ -1,9 +1,9 @@
 use super::output::{
-    emit_ledger_sync, emit_live_fill, emit_reconciliation_snapshot_error,
-    emit_reconciliation_state, emit_startup_rejected, emit_stop_loss_triggered,
+    emit_live_fill, emit_reconciliation_snapshot_error, emit_reconciliation_state,
+    emit_stop_loss_triggered,
 };
 use super::*;
-use standx_sdk::order_response::{OrderCommandSender, OrderResponse, OrderResponseHealth};
+use standx_sdk::order_response::OrderResponse;
 
 fn take_cleanup_effect(
     runtime_state: &mut MakerState,
@@ -569,57 +569,6 @@ async fn accounting_invariant_exit(
     Some(MakerExit::AccountingInvariant(detail))
 }
 
-/// Reject alert thresholds that silently defeat the guard they configure.
-///
-/// A negative threshold turns the alert off without warning, and a percentage
-/// above 100 can never fire; both leave an operator believing a protection is
-/// armed when it is not.
-pub(super) fn validate_alert_thresholds(
-    alert_loss: f64,
-    alert_inventory_pct: f64,
-    alert_position_change_pct: f64,
-    alert_uptime: f64,
-) -> Result<()> {
-    if alert_loss < 0.0 {
-        return Err(anyhow::anyhow!("--alert-loss must be >= 0"));
-    }
-    if !(0.0..=100.0).contains(&alert_inventory_pct) {
-        return Err(anyhow::anyhow!("--alert-inventory-pct must be 0..=100"));
-    }
-    if !(0.0..=100.0).contains(&alert_position_change_pct) {
-        return Err(anyhow::anyhow!(
-            "--alert-position-change-pct must be 0..=100"
-        ));
-    }
-    if alert_uptime < 0.0 {
-        return Err(anyhow::anyhow!("--alert-uptime must be >= 0"));
-    }
-    Ok(())
-}
-
-/// Everything that exists exactly when the maker runs `--live`: the
-/// order-response stream, the authenticated account stream, the projection
-/// they feed, and the REST polling schedule for account floors.
-///
-/// `run_maker` holds a single `Option<LiveSession>` that is `Some` iff
-/// `args.live`, so live-only code paths bind the session once and use plain
-/// fields instead of re-proving "live implies initialized" per use. Reconnect
-/// paths replace the relevant fields in place (aborting the old task handle
-/// first); the session itself stays `Some` for the whole run.
-struct LiveSession {
-    order_responses: tokio::sync::mpsc::Receiver<OrderResponse>,
-    order_commands: OrderCommandSender,
-    order_response_health: OrderResponseHealth,
-    order_response_handle: tokio::task::JoinHandle<()>,
-    account_events: tokio::sync::mpsc::Receiver<AccountEvent>,
-    account_stream_health: AccountStreamHealth,
-    account_stream_handle: tokio::task::JoinHandle<()>,
-    /// Bumped on every account-stream reconnect; stamps projection generations.
-    account_stream_epoch: u64,
-    projection: MakerAccountProjection,
-    account_poll: LiveAccountPollState,
-}
-
 struct ShutdownReport<'a> {
     live: bool,
     output_format: OutputFormat,
@@ -807,427 +756,25 @@ async fn shutdown_report(report: ShutdownReport<'_>) -> Result<()> {
     }
 }
 
-fn new_maker_rest_client() -> Result<StandXClient> {
-    let client = StandXClient::new()?;
-    debug_assert!(client.session_id().is_none());
-    Ok(client)
-}
-
 pub(super) async fn run_maker(
     symbol: String,
     args: MakerRunArgs,
     output_format: OutputFormat,
 ) -> Result<()> {
-    let order_session_id = args.live.then(|| uuid::Uuid::new_v4().to_string());
-    let run_uuid = uuid::Uuid::new_v4().simple().to_string();
-    let run_order_prefix = format!("{}{}-", MAKER_CL_ORD_ID_PREFIX, &run_uuid[..12]);
-    let mut starting_position = 0.0_f64;
-    let mut baseline_mark = 0.0_f64;
-    let mut session_started_at = chrono::Utc::now().timestamp();
-
-    if let Some(after) = args.controlled_disconnect_after {
-        if !args.live {
-            return Err(anyhow::anyhow!(
-                "--controlled-disconnect-after requires --live"
-            ));
-        }
-        if after == 0 || after > 60 {
-            return Err(anyhow::anyhow!(
-                "--controlled-disconnect-after must be between 1 and 60 seconds"
-            ));
-        }
-    }
-    if args.order_response_reconnect_attempts > 10 {
-        return Err(anyhow::anyhow!(
-            "--order-response-reconnect-attempts must be between 0 and 10"
-        ));
-    }
-    if args.order_response_reconnect_attempts > 0
-        && !(1..=60).contains(&args.order_response_reconnect_backoff)
-    {
-        return Err(anyhow::anyhow!(
-            "--order-response-reconnect-backoff must be between 1 and 60 seconds when reconnect is enabled"
-        ));
-    }
-    if args.account_stream_reconnect_attempts > 10 {
-        return Err(anyhow::anyhow!(
-            "--account-stream-reconnect-attempts must be between 0 and 10"
-        ));
-    }
-    if args.account_stream_reconnect_attempts > 0
-        && !(1..=60).contains(&args.account_stream_reconnect_backoff)
-    {
-        return Err(anyhow::anyhow!(
-            "--account-stream-reconnect-backoff must be between 1 and 60 seconds when reconnect is enabled"
-        ));
-    }
-    if !(1..=100).contains(&args.recovery_incidents_per_window) {
-        return Err(anyhow::anyhow!(
-            "--recovery-incidents-per-window must be between 1 and 100"
-        ));
-    }
-    if !(60..=86_400).contains(&args.recovery_window_secs) {
-        return Err(anyhow::anyhow!(
-            "--recovery-window-secs must be between 60 and 86400"
-        ));
-    }
-    // REST reads, audits, and fail-safe cleanup must stay outside the
-    // order-response session. Attaching x-session-id to REST cancellation
-    // would route its asynchronous response into the command response stream,
-    // where it has no projection request entry and would look uncorrelated.
-    let client = new_maker_rest_client()?;
-
-    // ---- Startup: symbol metadata + invariants (fail fast) ----
-    let infos = client.get_symbol_info().await?;
-    let info = infos
-        .iter()
-        .find(|i| i.symbol.eq_ignore_ascii_case(&symbol))
-        .ok_or_else(|| {
-            anyhow::anyhow!(
-                "Unknown symbol '{}'. Available: {}",
-                symbol,
-                infos
-                    .iter()
-                    .map(|i| i.symbol.as_str())
-                    .collect::<Vec<_>>()
-                    .join(", ")
-            )
-        })?;
-    if info.status != "trading" {
-        return Err(anyhow::anyhow!(
-            "Symbol {} is not trading (status: {})",
-            info.symbol,
-            info.status
-        ));
-    }
-    let symbol = info.symbol.clone(); // canonical casing
-
-    let min_order_qty: f64 = info.min_order_qty.parse().map_err(|_| {
-        anyhow::anyhow!(
-            "unparseable min_order_qty '{}' for {} from venue symbol info",
-            info.min_order_qty,
-            info.symbol
-        )
-    })?;
-    let cfg = MakerConfig {
-        spread_bps: args.spread_bps,
-        band_bps: args.band_bps,
-        level_step_bps: args.level_step_bps,
-        refresh_bps: args.refresh_bps,
-        levels: args.levels.max(1),
-        size: args.size,
-        max_position: args.max_position,
-        skew_bps: args.skew_bps,
-        price_decimals: info.price_tick_decimals,
-        qty_decimals: info.qty_tick_decimals,
-        min_order_qty,
-    };
-
-    if cfg.spread_bps <= 0.0 {
-        return Err(anyhow::anyhow!("--spread-bps must be > 0"));
-    }
-    if cfg.skew_bps < 0.0 {
-        return Err(anyhow::anyhow!("--skew-bps must be >= 0"));
-    }
-    if !(0.0..=100.0).contains(&args.inventory_exit_pct) || args.inventory_exit_qty < 0.0 {
-        return Err(anyhow::anyhow!(
-            "--inventory-exit-pct must be 0..=100 and --inventory-exit-qty must be >= 0"
-        ));
-    }
-    if (args.inventory_exit_pct > 0.0) != (args.inventory_exit_qty > 0.0) {
-        return Err(anyhow::anyhow!(
-            "active inventory exit requires both --inventory-exit-pct and --inventory-exit-qty"
-        ));
-    }
-    validate_alert_thresholds(
-        args.alert_loss,
-        args.alert_inventory_pct,
-        args.alert_position_change_pct,
-        args.alert_uptime,
-    )?;
-    if cfg.band_bps <= cfg.spread_bps {
-        return Err(anyhow::anyhow!(
-            "--band-bps ({}) must be greater than --spread-bps ({}): quotes clamped to the band edge would sit exactly at the boundary",
-            cfg.band_bps,
-            cfg.spread_bps
-        ));
-    }
-    let rounded_size = maker::round_to_decimals(cfg.size, cfg.qty_decimals);
-    if rounded_size < cfg.min_order_qty || rounded_size <= 0.0 {
-        return Err(anyhow::anyhow!(
-            "--size {} (rounded to {} at {} decimals) is below min order qty {} for {}",
-            cfg.size,
-            rounded_size,
-            cfg.qty_decimals,
-            cfg.min_order_qty,
-            symbol
-        ));
-    }
-    if cfg.refresh_bps >= cfg.spread_bps {
-        eprintln!(
-            "⚠️  --refresh-bps ({}) >= --spread-bps ({}): quotes will be held through large drifts",
-            cfg.refresh_bps, cfg.spread_bps
-        );
-    }
-    if cfg.levels > 1
-        && cfg.spread_bps + (cfg.levels - 1) as f64 * cfg.level_step_bps >= cfg.band_bps
-    {
-        eprintln!("⚠️  outer quote levels exceed the band and will be clamped/collapsed");
-    }
-
-    let notifier = MakerNotifier::new(
-        output_format,
-        args.alert_webhook.clone(),
-        args.alert_webhook_format,
-    );
-
-    // Half a qty tick: the adoption/mismatch tolerance used throughout the run.
-    let qty_tolerance = 10_f64.powi(-(cfg.qty_decimals as i32)) / 2.0;
-
-    // ---- Live gating & clean start ----
-    // `order_session_id` is `Some` iff `args.live`, so this block is the live
-    // startup path; it either fails fast or yields a complete `LiveSession`.
-    let mut live_session: Option<LiveSession> = None;
-    if let Some(order_session_id) = order_session_id.as_deref() {
-        if std::env::var(LIVE_MAKER_ENV).ok().as_deref() != Some("1") {
-            return Err(anyhow::anyhow!(
-                "live mode not yet enabled: it has not been supervised-tested against production. Set {}=1 to unlock (at your own risk).",
-                LIVE_MAKER_ENV
-            ));
-        }
-        // A live run with no push channel is how #220 happens: if the process
-        // dies (SIGKILL/OOM/panic/host down) nobody is notified and resting
-        // orders are left on the venue. Refuse to start live without a
-        // webhook, and refuse if the webhook can never fire because every
-        // alert threshold is disabled.
-        if args.alert_webhook.is_none() {
-            return Err(anyhow::anyhow!(
-                "live mode requires --alert-webhook so the maker can push risk/stop notifications; refusing to run live with no push channel"
-            ));
-        }
-        if args.alert_loss <= 0.0
-            && args.alert_inventory_pct <= 0.0
-            && args.alert_position_change_pct <= 0.0
-            && args.alert_uptime <= 0.0
-            && args.alert_equity_below <= 0.0
-            && args.alert_margin_below <= 0.0
-        {
-            return Err(anyhow::anyhow!(
-                "live mode requires at least one alert threshold; all maker and account thresholds are 0 so the webhook would never fire"
-            ));
-        }
-        let creds = Credentials::load()?;
-        if creds.is_expired() {
-            return Err(anyhow::anyhow!(
-                "Credentials expired. Run 'standx auth login' first."
-            ));
-        }
-        if creds.private_key.is_empty() {
-            return Err(anyhow::anyhow!(
-                "Live mode requires a private key for order signing. Run 'standx auth login' with --private-key."
-            ));
-        }
-        let open_orders = client.get_open_orders(Some(&symbol)).await?;
-        let manual_orders = open_orders
-            .iter()
-            .filter(|order| !is_maker_order(order))
-            .count();
-        if manual_orders > 0 {
-            eprintln!(
-                "ℹ️  preserving {} manually-managed order(s) on {}; only {} orders are managed",
-                manual_orders, symbol, MAKER_CL_ORD_ID_PREFIX
-            );
-        }
-        // Clean only leftover orders owned by this maker. Manual/API orders
-        // are not part of the strategy's reconciliation state and must never
-        // be adopted or cancelled as stale.
-        cancel_maker_orders_with_retry(&client, &symbol, 3, output_format).await?;
-
-        // Establish the session ledger boundary before any new order can be
-        // submitted. Existing inventory is adopted at the current mark, so
-        // maker-session PnL starts at zero while account upnl remains intact.
-        let history_to = chrono::Utc::now().timestamp();
-        let history_from = history_to.saturating_sub(LEDGER_HISTORY_WINDOW_SECS);
-        let (positions, startup_market, filled_orders, historical_trades, balance) = tokio::join!(
-            client.get_positions(Some(&symbol)),
-            market_snapshot(&client, &symbol, None),
-            client.get_order_history(Some(&symbol), Some(ORDER_HISTORY_LIMIT)),
-            client.get_user_trades(
-                &symbol,
-                history_from,
-                history_to,
-                Some(TRADE_LOOKBACK_LIMIT)
-            ),
-            client.get_balance(),
-        );
-        let positions = positions?;
-        let (mark, _, _, _, _) = startup_market?;
-        let filled_orders = filled_orders?;
-        let historical_trades = historical_trades?;
-        let balance = balance?;
-        starting_position = position_for_symbol(&positions, &symbol)?;
-        baseline_mark = mark;
-        let account_poll = LiveAccountPollState::new(balance, std::time::Instant::now());
-
-        let historical_order_ids = filled_orders
-            .iter()
-            .filter(|order| is_maker_order(order))
-            .map(|order| {
-                order.id.parse::<u64>().map_err(|_| {
-                    anyhow::anyhow!(
-                        "historical maker order has non-integer exchange ID '{}'",
-                        order.id
-                    )
-                })
-            })
-            .collect::<Result<HashSet<_>>>()?;
-        let historical_maker_orders = historical_order_ids.len();
-        let historical_maker_trades = historical_trades
-            .iter()
-            .filter(|trade| {
-                trade
-                    .order_id
-                    .is_some_and(|order_id| historical_order_ids.contains(&order_id))
-            })
-            .count();
-
-        if !maker::position_within_limit(starting_position, cfg.max_position, cfg.qty_decimals) {
-            emit_startup_rejected(output_format, &symbol, starting_position, cfg.max_position);
-            let message = format!(
-                "starting position {starting_position:+.8} exceeds max_position {:.8}",
-                cfg.max_position
-            );
-            notifier
-                .risk(
-                    RiskNotice {
-                        kind: "startup_position_limit",
-                        severity: "critical",
-                        event: "startup_rejected",
-                        message: &message,
-                        symbol: &symbol,
-                        cycle: 0,
-                        position_before: None,
-                        position_after: Some(starting_position),
-                        expected: None,
-                        observed: Some(starting_position),
-                    },
-                    true,
-                )
-                .await;
-            return Err(anyhow::anyhow!(
-                "starting position {:+.8} exceeds max_position {:.8}",
-                starting_position,
-                cfg.max_position
-            ));
-        }
-        session_started_at = chrono::Utc::now().timestamp();
-
-        // Authenticated account state is a hard live dependency. Connect it
-        // before order-response readiness, then require a second REST
-        // snapshot so events buffered during authentication cannot create an
-        // unobserved startup gap.
-        let account_stream_epoch = 1_u64;
-        let account_stream = AccountStream::new(account_stream_epoch)?;
-        let (account_events, account_stream_health, account_stream_handle) = account_stream
-            .connect(&[
-                AccountChannel::Order,
-                AccountChannel::Position,
-                AccountChannel::Trade,
-                AccountChannel::Balance,
-            ])
-            .await?;
-        let post_auth_positions = client.get_positions(Some(&symbol)).await?;
-        let post_auth_position = position_for_symbol(&post_auth_positions, &symbol)?;
-        if (post_auth_position - starting_position).abs() > qty_tolerance {
-            account_stream_handle.abort();
-            notifier
-                .risk(
-                    RiskNotice {
-                        kind: "position_reconciliation",
-                        severity: "critical",
-                        event: "startup_sync_failed",
-                        message: "position changed while the account stream was authenticating",
-                        symbol: &symbol,
-                        cycle: 0,
-                        position_before: Some(starting_position),
-                        position_after: Some(post_auth_position),
-                        expected: Some(starting_position),
-                        observed: Some(post_auth_position),
-                    },
-                    true,
-                )
-                .await;
-            return Err(anyhow::anyhow!(
-                "position changed while account stream was authenticating: baseline {starting_position:+.8}, snapshot {post_auth_position:+.8}"
-            ));
-        }
-        let stream = OrderResponseStream::new(order_session_id)?;
-        let (order_commands, order_responses, order_response_health, order_response_handle) =
-            stream.connect().await?;
-        if let Some(after) = args.controlled_disconnect_after {
-            let health_for_fault = order_response_health.clone();
-            let abort = order_response_handle.abort_handle();
-            tokio::spawn(async move {
-                tokio::time::sleep(Duration::from_secs(after)).await;
-                // Aborting drops the local WebSocket halves; set health first
-                // so the maker exits through its existing fail-safe path even
-                // if the runtime delays observing the socket close.
-                health_for_fault.mark_unhealthy(format!(
-                    "controlled fault injection closed the order-response stream after {after}s"
-                ));
-                abort.abort();
-            });
-            eprintln!(
-                "⚠️ controlled fault injection armed: closing order-response stream after {after}s"
-            );
-        }
-        live_session = Some(LiveSession {
-            order_responses,
-            order_commands,
-            order_response_health,
-            order_response_handle,
-            account_events,
-            account_stream_health,
-            account_stream_handle,
-            account_stream_epoch,
-            projection: MakerAccountProjection::new(
-                account_stream_epoch,
-                run_order_prefix.clone(),
-                starting_position,
-                cfg.price_tick() / 2.0,
-                qty_tolerance,
-            ),
-            account_poll,
-        });
-        emit_ledger_sync(
-            output_format,
-            &symbol,
-            starting_position,
-            baseline_mark,
-            historical_maker_orders,
-            historical_maker_trades,
-        );
-        if starting_position.abs() > qty_tolerance {
-            let message = format!("adopted non-zero starting inventory {starting_position:+.8}");
-            notifier
-                .risk(
-                    RiskNotice {
-                        kind: "inventory_adopted",
-                        severity: "warning",
-                        event: "startup",
-                        message: &message,
-                        symbol: &symbol,
-                        cycle: 0,
-                        position_before: Some(0.0),
-                        position_after: Some(starting_position),
-                        expected: Some(starting_position),
-                        observed: Some(starting_position),
-                    },
-                    false,
-                )
-                .await;
-        }
-    }
+    // Validate args, resolve config, and (in live mode) run the clean-start
+    // handshake into a fully-initialized session before the quoting loop.
+    let MakerStartup {
+        client,
+        cfg,
+        symbol,
+        notifier,
+        qty_tolerance,
+        run_order_prefix,
+        starting_position,
+        baseline_mark,
+        session_started_at,
+        mut live_session,
+    } = run_startup(symbol, &args, output_format).await?;
 
     let mode = if args.live { "LIVE" } else { "PAPER" };
     if output_format == OutputFormat::Table {
